feat: add capability to manually trigger a fresh rechunking plan
Over time the layer plan due to reuse of the previous build's rechunking
plan might become undesirable/inefficient or due to issues in rpm-ostree
itself might force us to reset the layer plan where we do not consume
our previous build.

+++ Justfile
@@ -275,7 +275,7 @@ build-pipeline image="aurora" tag="latest" flavor="main" kernel_pin="":
 # Rechunk Image
 [group('Image')]
 [private]
-rechunk $image="aurora" $tag="latest" $flavor="main" ghcr="0" pipeline="0":
+rechunk $image="aurora" $tag="latest" $flavor="main" ghcr="0" pipeline="0" fresh_rechunk="0":
     #!/usr/bin/bash
 
     set -eoux pipefail
@@ -292,7 +292,7 @@ rechunk $image="aurora" $tag="latest" $flavor="main" ghcr="0" pipeline="0":
       {{ just }} load-rootful "${image}" "${tag}" "${flavor}"
     fi
 
-    if [[ "{{ ghcr }}" == "1" ]]; then
+    if [[ "{{ ghcr }}" == "1" && "{{ fresh_rechunk }}" == "0" ]]; then
       # TODO: Replace this with --previous-build in rpm-ostree 2026.1+
       # so we don't have to pull an old image anymore
       PREVIOUS_BUILD=ghcr.io/{{ repo_organization }}/"${image_name}":"${DEFAULT_TAG}"
@@ -321,7 +321,7 @@ rechunk $image="aurora" $tag="latest" $flavor="main" ghcr="0" pipeline="0":
         --output containers-storage:${CHUNKED_IMAGE}
 
     # Rename image from ghcr... to localhost... and discard PREVIOUS_BUILD
-    if [[ "{{ ghcr }}" == "1" ]]; then
+    if [[ "{{ ghcr }}" == "1" && "{{ fresh_rechunk }}" == "0" ]]; then
       ${SUDOIF} ${PODMAN} tag ${CHUNKED_IMAGE} "localhost/"${image_name}":"${tag}""
       ${SUDOIF} ${PODMAN} image rm -f ${CHUNKED_IMAGE}
     fi
